+++ src/listener.js
@@ -99,20 +99,16 @@
             true);
     }
 
-    OCA.Onlyoffice.onDocumentReady = function (documentType) {
-        if (documentType === "word"
-            || documentType === "cell"
-            || documentType === "slide") {
-            if (OCA.Onlyoffice.bindVersionClick) {
-                OCA.Onlyoffice.bindVersionClick();
-            }
-        } else if (OCA.Onlyoffice.unbindVersionClick) {
-            OCA.Onlyoffice.unbindVersionClick();
-        }
+    OCA.Onlyoffice.onDocumentReady = function () {
+        OCA.Onlyoffice.setViewport();
     };
 
     OCA.Onlyoffice.changeFavicon = function (favicon) {
         $('link[rel="icon"]').attr("href", favicon);
+    };
+
+    OCA.Onlyoffice.setViewport = function() {
+        document.querySelector('meta[name="viewport"]').setAttribute("content","width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=0");
     };
 
     OCA.Onlyoffice.onShowMessage = function (messageObj) {