+++ src/main.js
@@ -113,7 +113,7 @@ import NewDocxfSvg from "!!raw-loader!../img/new-docxf.svg";
                 callback(response);
 
                 if (open) {
-                    OCA.Onlyoffice.OpenEditor(response.id, dir, response.name, 0, winEditor);
+                    OCA.Onlyoffice.OpenEditor(response.id, dir, response.name, winEditor);
 
                     OCA.Onlyoffice.context = {
                         fileName: response.name,
@@ -126,7 +126,7 @@ import NewDocxfSvg from "!!raw-loader!../img/new-docxf.svg";
         );
     };
 
-    OCA.Onlyoffice.OpenEditor = function (fileId, fileDir, fileName, version, winEditor) {
+    OCA.Onlyoffice.OpenEditor = function (fileId, fileDir, fileName, winEditor) {
         var filePath = "";
         if (fileName) {
             filePath = fileDir.replace(new RegExp("\/$"), "") + "/" + fileName;
@@ -143,10 +143,6 @@ import NewDocxfSvg from "!!raw-loader!../img/new-docxf.svg";
                     shareToken: encodeURIComponent($("#sharingToken").val()),
                     fileId: fileId
                 });
-        }
-
-        if (version > 0) {
-            url += "&version=" + version;
         }
 
         if (winEditor && winEditor.location) {
@@ -188,8 +184,6 @@ import NewDocxfSvg from "!!raw-loader!../img/new-docxf.svg";
             window.history.pushState(null, null, url);
             OCA.Onlyoffice.folderUrl = null;
         }
-
-        OCA.Onlyoffice.bindVersionClick();
     };
 
     OCA.Onlyoffice.OpenShareDialog = function () {
@@ -218,7 +212,7 @@ import NewDocxfSvg from "!!raw-loader!../img/new-docxf.svg";
         var fileId = context.fileId || context.$file && context.$file[0].dataset.id || fileInfoModel.id;
         var winEditor = !fileInfoModel && !OCA.Onlyoffice.setting.sameTab ? document : null;
 
-        OCA.Onlyoffice.OpenEditor(fileId, context.dir, fileName, 0, winEditor);
+        OCA.Onlyoffice.OpenEditor(fileId, context.dir, fileName, winEditor);
 
         OCA.Onlyoffice.context = context;
         OCA.Onlyoffice.context.fileName = fileName;
@@ -755,49 +749,6 @@ import NewDocxfSvg from "!!raw-loader!../img/new-docxf.svg";
         return extension;
     }
 
-    OCA.Onlyoffice.openVersion = function (fileId, version) {
-        if (OCA.Onlyoffice.frameSelector) {
-            $(OCA.Onlyoffice.frameSelector)[0].contentWindow.OCA.Onlyoffice.onRequestHistory(version);
-            return;
-        }
-
-        OCA.Onlyoffice.OpenEditor(fileId, "", "", version)
-    };
-
-    OCA.Onlyoffice.bindVersionClick = function () {
-        OCA.Onlyoffice.unbindVersionClick();
-        $(document).on("click.onlyoffice-version", "#versionsTabView .downloadVersion", function() {
-            var ext = $(this).attr("download").split(".").pop();
-            if (!OCA.Onlyoffice.setting.formats[ext]
-                || !OCA.Onlyoffice.setting.formats[ext].def) {
-                return true;
-            }
-
-            var versionNodes = $("#versionsTabView ul.versions>li");
-            var versionNode = $(this).closest("#versionsTabView ul.versions>li")[0];
-
-            var href = $(this).attr("href");
-            var search = new RegExp("\/versions\/(\\d+)\/\\d+$");
-            var result = search.exec(href);
-            if (result && result.length > 1) {
-                var fileId = result[1];
-            }
-            if (!fileId) {
-                return true;
-            }
-
-            var versionNum = versionNodes.length - $.inArray(versionNode, versionNodes);
-
-            OCA.Onlyoffice.openVersion(fileId, versionNum);
-
-            return false;
-        });
-    };
-
-    OCA.Onlyoffice.unbindVersionClick = function() {
-        $(document).off("click.onlyoffice-version", "#versionsTabView .downloadVersion");
-    }
-
     var initPage = function () {
         if ($("#isPublic").val() === "1" && $("#mimetype").val() !== "httpd/unix-directory") {
             //file by shared link
@@ -839,8 +790,6 @@ import NewDocxfSvg from "!!raw-loader!../img/new-docxf.svg";
             OCA.Onlyoffice.registerNewFileMenu();
 
             OCA.Onlyoffice.registerAction();
-
-            OCA.Onlyoffice.bindVersionClick();
         }
     };