+++ src/share.js
@@ -288,7 +288,7 @@ import AppDarkSvg from "!!raw-loader!../img/app-dark.svg";
                 attributes.push({
                     checked: modifyFilter,
                     extra: OCA.Onlyoffice.Permissions.ModifyFilter,
-                    label: t(OCA.Onlyoffice.AppName, "Custom filter")
+                    label: t(OCA.Onlyoffice.AppName, "Global filter")
                 });
             }