+++ templates/settings.php
@@ -110,7 +110,7 @@ if ($_["tagsEnabled"]) {
     <p>
         <input type="checkbox" class="checkbox" id="onlyofficeGroups"
             <?php if (count($_["limitGroups"]) > 0) { ?>checked="checked"<?php } ?> />
-        <label for="onlyofficeGroups"><?php p($l->t("Restrict access to editors to following groups")) ?></label>
+        <label for="onlyofficeGroups"><?php p($l->t("Allow the following groups to access the editors")) ?></label>
         <input type="hidden" id="onlyofficeLimitGroups" value="<?php p(implode("|", $_["limitGroups"])) ?>" style="display: block" />
     </p>