+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - offline viewer for share link
 - updatable list of supported formats
 - filling pdf instead oform
+- fixed mobile editor size
+- version opening from right tab is no longer supported
 - compatible with Nextcloud 28
 - agpl v3 license
 
@@ -21,6 +23,7 @@
 - setting for disable editors cron check
 - Arabic and Serbian empty file templates
 - list of users to protect ranges of cells
+- support of user avatar in editor
 
 ## 8.2.4
 ## Changed